DCT Hash
The reference for the DCT hash appears to describe usage only of
a small section of the DCT coefficients of the whole image, rather
than select low-frequency coefficients from application of DCT to
the image in 8x8 blocks, as I would've expected.

Checking the implementation in pHash seems to confirm this. However,
the pHash ph_dct_imagehash function appears to return different hash
values for cases where I would expect this to fail, and in general
the hash values I get from that function are very different from
those my own implementation outputs. Initial experiments had me wondering
if this was a byte-order issue, but it appears not, and I'm struggling
to find the point of divergence.