Stash of work on MH hash
The bulk of the dev time here was spent implementing histogram
equalisation, which now works for the required input image types.

The next bit was building the MH kernel function, which I did by
reference to the GetMHKernel function from pHash. Some other
references might help me be sure that the output kernel is correct.

It now appears that the imfilter_LoG I had thought would stand in
for get_correlate does not. I need to implement the `correlate`
function of CImg.

After that, I get to dive into the madness that unrolls the
transformed image into a hash value.